+++ pkg/monitoring/alloy/service.go
@@ -13,6 +13,7 @@ import (
 	"github.com/pkg/errors"
 
 	"github.com/giantswarm/observability-operator/pkg/common"
+	commonmonitoring "github.com/giantswarm/observability-operator/pkg/common/monitoring"
 	"github.com/giantswarm/observability-operator/pkg/common/organization"
 	"github.com/giantswarm/observability-operator/pkg/common/password"
 	"github.com/giantswarm/observability-operator/pkg/monitoring"
@@ -35,9 +36,16 @@ func (a *Service) ReconcileCreate(ctx context.Context, cluster *clusterv1.Cluste
 	logger := log.FromContext(ctx)
 	logger.Info("alloy-service - ensuring alloy is configured")
 
+	// Get list of tenants
+	var tenants = []string{}
+	tenants, err := commonmonitoring.ListTenants(a.Client, ctx)
+	if err != nil {
+		return errors.WithStack(err)
+	}
+
 	configmap := ConfigMap(cluster)
-	_, err := controllerutil.CreateOrUpdate(ctx, a.Client, configmap, func() error {
-		data, err := a.GenerateAlloyMonitoringConfigMapData(ctx, configmap, cluster)
+	_, err = controllerutil.CreateOrUpdate(ctx, a.Client, configmap, func() error {
+		data, err := a.GenerateAlloyMonitoringConfigMapData(ctx, configmap, cluster, tenants)
 		if err != nil {
 			logger.Error(err, "alloy-service - failed to generate alloy monitoring configmap")
 			return errors.WithStack(err)